feat: add ExpressionExtensions_Nacos.cs
